Add support for ceph-proxy dataset
Add parsing of munged data presented by the ceph-proxy
charm in the ceph-public-address relation key; this
includes dropping the addition of the port to the values
added to the list of monitor hosts - this is not required
and the ceph-proxy may provide that data as part of the
ceph-public-address.

This change brings this interface inline with existing
ceph context support in the openstack charm helpers.

Change-Id: Id4d7b5df711c378763d5adb54269a3ede344dd45