import zuul job settings from project-config
This is a mechanically generated patch to complete step 1 of moving
the zuul job settings out of project-config and into each project
repository.

Because there will be a separate patch on each branch, the branch
specifiers for branch-specific jobs have been removed.

Because this patch is generated by a script, there may be some
cosmetic changes to the layout of the YAML file(s) as the contents are
normalized.

See the python3-first goal document for details:
https://governance.openstack.org/tc/goals/stein/python3-first.html

Remove --allow-unverified from pip, it's not supported with current pip.
Remove install of python-apt since it fails and is unneeded for the job.

Change-Id: I72ab95ec19b059ef83e628859cf16e663333308b
Story: #2002586
Task: #24317